refactor: activity list transaction
