+++ src/app/common/hooks/use-on-resize-listener.ts
@@ -1,0 +1,9 @@
+import { useEffect } from 'react';
+
+export function useOnResizeListener(callback: () => void) {
+  useEffect(() => {
+    const onResize = () => callback();
+    window.addEventListener('resize', onResize);
+    return () => window.removeEventListener('resize', onResize);
+  }, [callback]);
+}
