+++ src/app/features/increase-fee-drawer/components/increase-fee-form.tsx
@@ -10,7 +10,7 @@ import { microStxToStx, stacksValue, stxToMicroStx } from '@app/common/stacks-ut
 import { useRefreshAllAccountData } from '@app/common/hooks/account/use-refresh-all-account-data';
 import { useFeeSchema } from '@app/common/validation/use-fee-schema';
 import { Caption } from '@app/components/typography';
-import { TxItem } from '@app/components/tx-item';
+import { TransactionItem } from '@app/components/transaction/components/transaction-item';
 import { useRawDeserializedTxState, useRawTxIdState } from '@app/store/transactions/raw.hooks';
 import { useReplaceByFeeSubmitCallBack } from '@app/store/transactions/fees.hooks';
 import { useCurrentAccountAvailableStxBalance } from '@app/store/accounts/account.hooks';
@@ -66,7 +66,7 @@ export function IncreaseFeeForm(): JSX.Element | null {
     >
       {() => (
         <Stack spacing="extra-loose">
-          {tx && <TxItem position="relative" zIndex={99} transaction={tx} />}
+          {tx && <TransactionItem position="relative" transaction={tx} zIndex={99} />}
           <Stack spacing="base">
             <IncreaseFeeField currentFee={fee} />
             {stxBalance && (
